+++ library/src/main/java/com/github/techisfun/onelinecalendar/OneLineCalendarContract.java
@@ -1,7 +1,6 @@
 package com.github.techisfun.onelinecalendar;
 
-import android.support.v7.widget.RecyclerView;
-
+import androidx.recyclerview.widget.RecyclerView;
 import java.util.List;
 
 /**