+++ library/src/main/java/com/github/techisfun/onelinecalendar/OneLineCalendarPresenter.java
@@ -1,8 +1,8 @@
 package com.github.techisfun.onelinecalendar;
 
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
@@ -63,7 +63,7 @@ class OneLineCalendarPresenter implements OneLineCalendarContract.Presenter {
             private LinearLayoutManager mLayoutManager = null;
 
             @Override
-            public void onScrolled(RecyclerView recyclerView, int dx, int dy) {
+            public void onScrolled(@NonNull RecyclerView recyclerView, int dx, int dy) {
                 super.onScrolled(recyclerView, dx, dy);
 
                 if (mLayoutManager == null) {