+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # v0.2 [Dec 2019]
 - Updated to gradle plugin 3.5.3
 - Updated to target sdk 28 and min sdk 17
+- Updated to androidx.appcompat:appcompat:1.1.0
+- Updated to androidx.recyclerview:recyclerview:1.1.0
 
 * v0.1 Project started! (Tue Sep 12 16:30:33 2017)